+++ src/Sharers/FacebookSharer.php
@@ -10,24 +10,26 @@ namespace MichaelHall\LinkSharer\Sharers;
 
 use DataTypes\Interfaces\UrlInterface;
 use DataTypes\Url;
+use MichaelHall\LinkSharer\Sharers\Base\AbstractSharer;
 
 /**
  * Facebook sharer.
  *
  * @since 1.0.0
  */
-class FacebookSharer
+class FacebookSharer extends AbstractSharer
 {
     /**
      * Constructs a FacebookSharer.
      *
      * @since 1.0.0
      *
-     * @param UrlInterface $url The url.
+     * @param UrlInterface $url      The url.
+     * @param array        $hashtags The hashtags.
      */
-    public function __construct(UrlInterface $url)
+    public function __construct(UrlInterface $url, array $hashtags)
     {
-        $this->url = $url;
+        parent::__construct($url, '', $hashtags);
     }
 
     /**
@@ -39,23 +41,14 @@ class FacebookSharer
      */
     public function getShareUrl(): UrlInterface
     {
-        return Url::parse('https://www.facebook.com/sharer/sharer.php?u=' . rawurlencode($this->url->__toString()));
-    }
+        $parts = [];
 
-    /**
-     * Returns the share url as a string.
-     *
-     * @since 1.0.0
-     *
-     * @return string The share url as a string.
-     */
-    public function __toString(): string
-    {
-        return $this->getShareUrl()->__toString();
-    }
+        $parts[] = 'u=' . rawurlencode($this->getUrl()->__toString());
 
-    /**
-     * @var Url My url.
-     */
-    private $url;
+        if (count($this->getHashtags()) > 0) {
+            $parts[] = 'hashtag=' . rawurlencode('#' . $this->getHashtags()[0]);
+        }
+
+        return Url::parse('https://www.facebook.com/sharer/sharer.php?' . implode('&', $parts));
+    }
 }