+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## [2.0.0]
+## [2.1.0] - 2018-10-04
+- Updated Twitter sharer url.
+- Added hashtag functionality to Facebook sharer.
+- Added Reddit sharer.
+
+## [2.0.0] - 2017-12-26
 - Changed required PHP version to >= 7.1.
 - Changed dependency to DataTypes 2.x.
 
@@ -8,5 +13,6 @@
 ## 1.0.0 - 2017-06-05
 - First stable revision.
 
-[2.0.0]: https://github.com/themichaelhall/link-sharer/compare/v1.1.1...v2.0.0
+[2.1.0]: https://github.com/themichaelhall/link-sharer/compare/v2.0.0...v2.1.0
+[2.0.0]: https://github.com/themichaelhall/link-sharer/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/themichaelhall/link-sharer/compare/v1.0.0...v1.1.0